requirements.txt
# Core FastFusion/Diffusers stack
torch>=2.1.0
diffusers @ git+https://github.com/huggingface/diffusers
transformers
accelerate
safetensors

# API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
requests>=2.31.0

# Image handling
pillow>=10.2.0
numpy>=1.26.0

# Utilities
huggingface-hub>=0.21.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Health checks
httpx>=0.27.0
hf_transfer

# SAM3 and image processing
opencv-python>=4.8.0
scipy>=1.11.0